Split out sexp tests from fileformats testers
